+++ WalletWasabi.Fluent/ViewModels/Login/PasswordFinder/ContainsNumbersViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System.Reactive;
 using ReactiveUI;
 using WalletWasabi.Fluent.ViewModels.Navigation;
+using WalletWasabi.Wallets.PasswordFinder;
 
 namespace WalletWasabi.Fluent.ViewModels.Login.PasswordFinder
 {
