+++ WalletWasabi.Fluent/ViewModels/Login/PasswordFinder/SelectCharsetViewModel.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Reactive.Linq;
 using ReactiveUI;
 using WalletWasabi.Fluent.ViewModels.Navigation;
+using WalletWasabi.Wallets.PasswordFinder;
 
 namespace WalletWasabi.Fluent.ViewModels.Login.PasswordFinder
 {
