+++ WalletWasabi/Wallets/PasswordFinder/PasswordFinderCharset.cs
@@ -1,0 +1,22 @@
+using System.ComponentModel;
+
+namespace WalletWasabi.Wallets.PasswordFinder
+{
+	public enum Charset
+	{
+		[Description("English")]
+		en,
+
+		[Description("Spanish")]
+		es,
+
+		[Description("Italian")]
+		it,
+
+		[Description("French")]
+		fr,
+
+		[Description("Portuguese")]
+		pt
+	}
+}
